Move ppoll_compat to its own module
